+++ jgclark.Dashboard/src/dataGenerationSearch.js
@@ -7,7 +7,7 @@
 import pluginJson from '../plugin.json'
 import { extendedSearch } from '../../jgclark.SearchExtensions/src/externalSearch'
 import { getSearchSettings } from '../../jgclark.SearchExtensions/src/searchHelpers'
-import type { noteAndLine, resultOutputV3Type, TSearchOptions } from '../../jgclark.SearchExtensions/src/searchHelpers'
+import type { noteAndLine, TSearchResultSet, TSearchOptions } from '../../jgclark.SearchExtensions/src/searchHelpers'
 import { WEBVIEW_WINDOW_ID } from './constants'
 import { savedSearch1 } from './demoData'
 import type { TDashboardSettings, TSection, TSectionItem } from './types'
@@ -97,7 +97,7 @@ export async function getSearchResults(searchTermsStr: string, config: TDashboar
     const maxInSection = config.maxItemsToShowInSection
 
     // Main search call to jgclark.SearchExtensions (NotePlan advanced/native syntax; NP 3.18.1+)
-    const searchResultSet: ?resultOutputV3Type = await extendedSearch(searchTermsStr, searchOptions)
+    const searchResultSet: ?TSearchResultSet = await extendedSearch(searchTermsStr, searchOptions)
     if (!searchResultSet) {
       logError('getSearchResults', 'extendedSearch returned null/undefined')
       return []
@@ -297,7 +297,7 @@ export async function getSavedSearchResults(
       // // TODO: ...
 
       // // Main search call to jgclark.SearchExtensions, that includes Perspective folder-level filtering, and item-defeating, but it doesn't cover ignoring certain sections within a note.
-      // const searchResultSet: resultOutputV3Type = await extendedSearch(extendedSearchTerms, searchOptions)
+      // const searchResultSet: TSearchResultSet = await extendedSearch(extendedSearchTerms, searchOptions)
       // const searchTermsRep = searchResultSet.searchTermsRepArr.join(' ')
       // const resultNALs: Array<noteAndLine> = searchResultSet.resultNoteAndLineArr
       // logDebug('getSavedSearchResults', `- found ${resultNALs.length} items from [${searchTermsRep}]`)