+++ jgclark.SearchExtensions/__tests__/nativeSearch.test.js
@@ -1,6 +1,6 @@
 /* global describe, expect, test, beforeAll */
 // @flow
-import * as s from '../src/NPExtendedSyntaxHelpers'
+import * as s from '../src/nativeSearch'
 import type { reducedFieldSet } from '../src/searchHelpers'
 import { Calendar, Clipboard, CommandBar, DataStore, Editor, NotePlan } from '@mocks/index'
 
@@ -14,26 +14,26 @@ beforeAll(() => {
   DataStore.settings['_logLevel'] = 'none' //change this to DEBUG to get more logging
 })
 
-describe('NPExtendedSyntaxHelpers.js tests', () => {
-  describe('getNonNegativeSearchTermsFromNPExtendedSyntax', () => {
+describe('nativeSearch.js tests', () => {
+  describe('getHighlightTermsFromNativeSearch', () => {
     test('should return empty array from empty string', () => {
-      const result = s.getNonNegativeSearchTermsFromNPExtendedSyntax('')
+      const result = s.getHighlightTermsFromNativeSearch('')
       expect(result).toEqual([])
     })
     test('should return empty array from only operators', () => {
-      const result = s.getNonNegativeSearchTermsFromNPExtendedSyntax('is:open')
+      const result = s.getHighlightTermsFromNativeSearch('is:open')
       expect(result).toEqual([])
     })
     test('should return positive terms from mixed input 1', () => {
-      const result = s.getNonNegativeSearchTermsFromNPExtendedSyntax('Holy Spirit -odd -is:open')
+      const result = s.getHighlightTermsFromNativeSearch('Holy Spirit -odd -is:open')
       expect(result).toEqual(['Holy','Spirit'])
     })
     test('should return array with all terms from mixed input 2 with joint OR', () => {
-      const result = s.getNonNegativeSearchTermsFromNPExtendedSyntax('Holy Spirit (odd OR even)')
+      const result = s.getHighlightTermsFromNativeSearch('Holy Spirit (odd OR even)')
       expect(result).toEqual(['Holy','Spirit', 'odd', 'even'])
     })
     test('should return empty array from mixed input 3 with joint negation', () => {
-      const result = s.getNonNegativeSearchTermsFromNPExtendedSyntax('Holy Spirit -(odd OR even)')
+      const result = s.getHighlightTermsFromNativeSearch('Holy Spirit -(odd OR even)')
       expect(result).toEqual(['Holy','Spirit'])
     })
   })